Some tweaks here and there.  Tided up comment in lilypondio.cpp. Attempted to get picthbend wheel working (and failed) and changed default pan value to central (for audio pan position).  Also added some RPN/NRPN convenience calls (untested).
Additionally fixed 703702 (Record button on toolbar doesn't work).


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@3921 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/AlsaDriver.cpp
@@ -1845,8 +1845,8 @@ AlsaDriver::processMidiOut(const MappedComposition &mC,
 
             case MappedEvent::MidiPitchBend:
                 {
-                    int value = ((*i)->getData1() << 7) |
-                                ((*i)->getData2() & 0x7F);
+                    int value = (((int)(*i)->getData1()) << 7) |
+                                (((int)(*i)->getData2()) & 0x7F);
 
                     // keep within -8192 to +8192
                     //
